test: add benchmarks
